requirements.txt
adium-theme-ubuntu==0.3.4
attrs==17.3.0
Automat==0.6.0
backports.ssl-match-hostname==3.4.0.2
cached-property==1.3.0
Cerberus==0.9.2
certifi==2017.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
crcmod==1.7
cryptography==1.2.3
Cython==0.27.3
django-filter==1.1.0
djangorestframework==3.7.3
docker-compose==1.8.0
docker-py==1.9.0
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.2
Eve==0.7.5
Events==0.2.2
ez-setup==0.9
Flask==0.12
Flask-PyMongo==0.5.1
funcsigs==0.4
functools32==3.2.3.post2
google-api-python-client==1.6.4
httplib2==0.10.3
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.5.1
logilab-common==1.1.0
Markdown==2.6.9
MarkupSafe==0.23
mock==1.3.0
ndg-httpsclient==0.4.0
numpy==1.13.3
oauth2client==4.1.2
pbr==1.8.0
pipreqs==0.4.9
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycurl==7.43.0
pymongo==3.6.0
pyOpenSSL==0.15.1
pystan==2.17.0.0
python-apt==1.1.0b1
pythoscope==0.4.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
simplejson==3.13.2
six==1.11.0
texttable==0.8.1
Twisted==17.9.0
unity-lens-photos==1.0
uritemplate==3.0.0
urllib3==1.22
websocket-client==0.18.0
Werkzeug==0.11.15
yarg==0.1.9
zope.interface==4.4.3
